Long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@6345 a1c6a512-1295-4272-9138-f99709370657

+++ apps/recorder/bmp.c
@@ -71,7 +71,7 @@ short readshort(void* value) {
     return bytes[0] | (bytes[1] << 8);
 }
 
-int readlong(void* value) {
+long readlong(void* value) {
     unsigned char* bytes = (unsigned char*) value;
     return bytes[0] | (bytes[1] << 8) | (bytes[2] << 16) | (bytes[3] << 24);
 }
